fix: revert delayed sheet registration
- Reverts delayed registration for action sheets
- Renames sheets.tsx to index.ts
- Removes LazySheets.tsx
