--- src/sheets/LazySheets.tsx
@@ -1,22 +1,0 @@
-import { lazy } from 'react';
-import { registerSheet } from 'react-native-actions-sheet';
-
-const QRScanner = lazy(() => import('../components/QRScanner.tsx'));
-const BackupPrompt = lazy(() => import('../components/BackupPrompt.tsx'));
-const ConfirmAuth = lazy(() => import('../components/ConfirmAuth.tsx'));
-const DeletePubky = lazy(() => import('../components/DeletePubky.tsx'));
-const AddPubky = lazy(() => import('../components/AddPubky.tsx'));
-const EditPubky = lazy(() => import('../components/EditPubky.tsx'));
-const SelectPubky = lazy(() => import('../components/SelectPubky.tsx'));
-const RecoveryPhrasePrompt = lazy(() => import('../components/RecoveryPhrasePrompt.tsx'));
-
-export const initializeSheets = (): void => {
-	registerSheet('camera', QRScanner as any);
-	registerSheet('backup-prompt', BackupPrompt as any);
-	registerSheet('recovery-phrase-prompt', RecoveryPhrasePrompt as any);
-	registerSheet('confirm-auth', ConfirmAuth as any);
-	registerSheet('delete-pubky', DeletePubky as any);
-	registerSheet('edit-pubky', EditPubky as any);
-	registerSheet('add-pubky', AddPubky as any);
-	registerSheet('select-pubky', SelectPubky as any);
-};
